requirements.txt
pyserial==3.5
jupyter==1.1.1